Release right before the first public CGAL Release. Many bugs in HTML converter fixed, including visit in Berlin and bug reports from Sven: \ref geaendert. \ccc in math mode. Gnu compiler warning in database.h. Single letter classes no longer indexed. -cgal_dir relaxed. If \ccHtmlNoClassIndex then no other member produces index entries. \ccEmptyParamter eliminated.

+++ Packages/Manual_tools/src/buffer.h
@@ -222,7 +222,7 @@ public:
 	*data = '\0';
     }
     void capitalize() {
-	for ( int i = 0; i < len; i++) 
+	for ( size_t i = 0; i < len; i++) 
 	    data[i] = toupper( data[i]);
     }
 };
